UPD: Refactor page rotation
